+++ Assets/04.Characters/Maincharactor(Player)/Mainchac_new.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 89f93233fec075d498040189f428c321
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
