+++ Assets/04.Characters/Maincharactor(Player)/PlayerScripts/Controll_Doll_JH.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0405895d17352c64f89e79d84095444a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
