+++ Assets/04.Characters/Maincharactor(Player)/Used.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 501490bc9208e124594c837a448e815d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
