--- Assets/BetaAssets/Jh9mintest/MAINCHAtest/NewPSD.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 501490bc9208e124594c837a448e815d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
