+++ Assets/BetaAssets/Jh9mintest/TestAp.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 99294f3d1c252f34e9723282dfa57619
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
